Add remove_from_trash to conversations_controller
add conversations#remove_from_trash action and corresponding
/conversations/:id/remove-from-trash route.
Expects the :id for the conversation to be removed the trash in dynamic
segment of the route.

Depending on what type of user is signed in: The conversations
show_to_column will be set to true.

next steps: add index and show actions to the conversations_controller.
refactor the set_conversation to ensure ownership.