Add show action to conversations_controller.rb
Create conversation.sh for testing.

The show actions route is /conversations/:id the the dynamic is segment
must be the id of the conversation. The show action will only show
the conversation if the signed in engineer or recruiter has ownership.

next steps: in conversations#update and conversations#show
return a 400 if the wrong account type is making the request